pyproject.toml
[tool.poetry]
name = "ppo"
description = "Reimplementation of PPO with Pytorch"
version = "0.1.0"
authors = [
    "Ying Zhang <ying.zh1992@gmail.com>"
]

[tool.poetry.dependencies]
python = "^3.8.5"
torch = "2.0.0"
torchvision = "0.15.1"
ipython = "8.12.0"
hydra-core = "1.3.2"
setuptools = "67.8.0"
wheel = "0.38.4"
opencv-python = "^4.7.0"
gymnasium = "0.28.1"

[tool.poetry.dev-dependencies]
pytest = "7.2.2"
tensorboard = "2.12.1"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
